python-sidecar/pyproject.toml
[project]
name = "aimbase-python-sidecar"
version = "0.3.0"
description = "Aimbase Python Sidecar - MCP Servers for RAG Pipeline, PII Safety, and Evaluation"
requires-python = ">=3.12"
dependencies = [
    # MCP Server
    "fastmcp>=2.0.0",
    # Document Parsing
    "unstructured[all-docs]>=0.16.0",
    # Semantic Chunking
    "llama-index-core>=0.12.0",
    "llama-index-embeddings-huggingface>=0.4.0",
    # Embeddings
    "sentence-transformers>=3.3.0",
    # Keyword Search
    "rank-bm25>=0.2.2",
    # Database
    "psycopg[binary]>=3.2.0",
    "pgvector>=0.3.0",
    # PII Detection (Sprint 16)
    "presidio-analyzer>=2.2.0",
    "presidio-anonymizer>=2.2.0",
    "spacy>=3.7.0",
    # Evaluation (Sprint 17)
    "ragas>=0.2.0",
    "deepeval>=2.0.0",
    "datasets>=3.0.0",
    # Utilities
    "numpy>=1.26.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
